reorganise extension into submodules
