+++ setup.py
@@ -1,3 +1,5 @@
+from pathlib import Path
+
 from pybind11.setup_helpers import Pybind11Extension, build_ext
 from setuptools import find_packages, setup
 from setuptools.command.build_ext import new_compiler
@@ -13,10 +15,11 @@ else:
 
 spcalext = Pybind11Extension(
     "spcal.lib.spcalext",
-    sources=["src/spcalext.cpp"],
+    sources=sorted(Path("src").glob("*.cpp")),
     cxx_std=cxx_std,
     extra_link_args=extra_link_args,
 )
+
 
 setup(
     packages=find_packages(include=["spcal", "spcal.*"]),
